Run prettier on the whole project

+++ api/src/storage/gridfs.js
@@ -8,7 +8,7 @@ async function getFile(cacheId, cacheHash, bucketName = 'fs') {
     const file = await collection.findOne({ _id: cacheId });
 
     if (file && file.metadata.hash === cacheHash) {
-        return { file, stream: bucket.openDownloadStream(cacheId) }
+        return { file, stream: bucket.openDownloadStream(cacheId) };
     }
 }
 
@@ -18,15 +18,12 @@ async function updateFile(cacheId, cacheHash, filename, mimeType, createStream, 
     const file = await collection.findOne({ _id: cacheId });
 
     if (!file || file.metadata.hash !== cacheHash) {
-        if (file)
-            await bucket.delete(cacheId);
+        if (file) await bucket.delete(cacheId);
 
         const readStream = await createStream();
-        const writeStream = bucket.openUploadStreamWithId(
-            cacheId,
-            filename,
-            { metadata: { hash: cacheHash, mimeType: mimeType } }
-        );
+        const writeStream = bucket.openUploadStreamWithId(cacheId, filename, {
+            metadata: { hash: cacheHash, mimeType: mimeType },
+        });
 
         await util.promisify(stream.pipeline)(readStream, writeStream);
     }
@@ -35,15 +32,21 @@ async function updateFile(cacheId, cacheHash, filename, mimeType, createStream, 
 async function deleteFiles(cacheIdPrefix, bucketName = 'fs') {
     const bucket = new GridFSBucket(database, { bucketName });
     const collection = database.collection(`${bucketName}.files`);
-    const files = await collection.find(
-        { _id: { $regex: new RegExp(`^${cacheIdPrefix}`) } },
-        { projection: { _id: 1 } }
-    ).toArray();
+    const files = await collection
+        .find({ _id: { $regex: new RegExp(`^${cacheIdPrefix}`) } }, { projection: { _id: 1 } })
+        .toArray();
 
     await Promise.all(files.map(file => bucket.delete(file._id)));
 }
 
-async function getGeneratedFile(sourceId, sourceHash, task, taskAttr, thumbnail = false, bucketName = 'fs') {
+async function getGeneratedFile(
+    sourceId,
+    sourceHash,
+    task,
+    taskAttr,
+    thumbnail = false,
+    bucketName = 'fs'
+) {
     let result;
 
     if (thumbnail) {
@@ -52,7 +55,14 @@ async function getGeneratedFile(sourceId, sourceHash, task, taskAttr, thumbnail 
         result = await getFile(thumbnailId, sourceHash, bucketName);
         if (!result) {
             // Make sure original is cached by recursing.
-            const original = await getGeneratedFile(sourceId, sourceHash, task, taskAttr, false, bucketName);
+            const original = await getGeneratedFile(
+                sourceId,
+                sourceHash,
+                task,
+                taskAttr,
+                false,
+                bucketName
+            );
             original.stream.destroy();
 
             const thumbnailJob = await queue.add(
@@ -65,8 +75,7 @@ async function getGeneratedFile(sourceId, sourceHash, task, taskAttr, thumbnail 
 
             result = getFile(thumbnailId, sourceHash, bucketName);
         }
-    }
-    else {
+    } else {
         result = await getFile(sourceId, sourceHash, bucketName);
         if (!result) {
             const documentJob = await queue.add(
